.vscode/requirements.txt
python-dotenv
openai
langchain
langchain-groq